# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Applyflow API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_v1_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Create Profile",
        "operationId": "create_profile_api_v1_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "profile"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_api_v1_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackgroundTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/retry": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Retry Task",
        "operationId": "retry_task_api_v1_tasks__task_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BackgroundTaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "task_type": {
            "$ref": "#/components/schemas/BackgroundTaskType"
          },
          "status": {
            "$ref": "#/components/schemas/BackgroundTaskStatus"
          },
          "provider_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Job Id"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts"
          },
          "can_retry": {
            "type": "boolean",
            "title": "Can Retry"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "next_retry_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Retry At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "user_id",
          "job_id",
          "task_type",
          "status",
          "provider_job_id",
          "payload",
          "result",
          "error_message",
          "attempt_count",
          "max_attempts",
          "can_retry",
          "created_at",
          "updated_at",
          "last_attempt_at",
          "next_retry_at",
          "completed_at"
        ],
        "title": "BackgroundTaskResponse"
      },
      "BackgroundTaskStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "BackgroundTaskStatus"
      },
...
      "BackgroundTaskType": {
        "type": "string",
        "enum": [
          "score_job",
          "generate_cover_letter"
        ],
        "title": "BackgroundTaskType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileCreateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "resume_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Text"
          },
          "context_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Text"
          },
          "resume_upload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadedProfileDocumentRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "context_upload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadedProfileDocumentRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfileCreateRequest"
      },
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "resume_text": {
            "type": "string",
            "title": "Resume Text"
          },
          "context_text": {
            "type": "string",
            "title": "Context Text"
          },
          "resume_version_id": {
            "type": "string",
            "title": "Resume Version Id"
          },
          "resume_version_number": {
            "type": "integer",
            "title": "Resume Version Number"
          },
          "context_version_id": {
            "type": "string",
            "title": "Context Version Id"
          },
          "context_version_number": {
            "type": "integer",
            "title": "Context Version Number"
          },
          "resume_source_file": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoredProfileDocumentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "context_source_file": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoredProfileDocumentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "display_name",
          "location",
          "resume_text",
          "context_text",
          "resume_version_id",
          "resume_version_number",
          "context_version_id",
          "context_version_number",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileResponse"
      },
      "ProfileUpdateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "resume_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Text"
          },
          "context_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Text"
          },
          "resume_upload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadedProfileDocumentRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "context_upload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadedProfileDocumentRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfileUpdateRequest"
      },
...
      "StoredProfileDocumentResponse": {
        "properties": {
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "file_name",
          "content_type",
          "size_bytes"
        ],
        "title": "StoredProfileDocumentResponse"
      },
      "TaskAcceptedResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "$ref": "#/components/schemas/BackgroundTaskStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskAcceptedResponse"
      },
      "UploadedProfileDocumentRequest": {
        "properties": {
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "content_base64": {
            "type": "string",
            "title": "Content Base64"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "file_name",
          "content_base64"
        ],
        "title": "UploadedProfileDocumentRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}